+++ assets/js/src/staircase.helper.js
@@ -1,41 +1,26 @@
-class staircase {
-  static add(id, type) {
-    let staircase = new StaircaseCore();
-    staircase.add(id, type);
+const staircase = {
+  add(id, type) {
+    new StaircaseCore().add(id, type);
+  },
+  delete(id) {
+    new StaircaseCore().delete(id);
+  },
+  rename(id, name) {
+    new StaircaseCore().rename(id, name);
+  },
+  select(id, callback = true) {
+    new StaircaseCore().select(id, callback);
+  },
+  deselect(callback = true) {
+    new StaircaseCore().deselect(callback);
+  },
+  open(id) {
+    new StaircaseCore().open(id);
+  },
+  close(id) {
+    new StaircaseCore().close(id);
+  },
+  refresh(id) {
+    new StaircaseCore().refresh(id);
   }
-
-  static delete(id) {
-    let staircase = new StaircaseCore();
-    staircase.delete(id);
-  }
-
-  static rename(id, name) {
-    let staircase = new StaircaseCore();
-    staircase.rename(id, name);
-  }
-
-  static select(id, callback = true) {
-    let staircase = new StaircaseCore();
-    staircase.select(id, callback);
-  }
-
-  static deselect(id, callback = true) {
-    let staircase = new StaircaseCore();
-    staircase.deselect(id, callback);
-  }
-
-  static open(id) {
-    let staircase = new StaircaseCore();
-    staircase.open(id);
-  }
-
-  static close(id) {
-    let staircase = new StaircaseCore();
-    staircase.close(id);
-  }
-
-  static refresh(id) {
-    let staircase = new StaircaseCore();
-    staircase.refresh(id);
-  }
-}
+};